# docs/openapi3.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Sample API
  description: A sample API to illustrate OpenAPI concepts
servers:
  - url: http://localhost:8080/
paths:
...
  /users/{email}:
    parameters:
      - name: email
        in: path
        description: email of user
        required: true
        schema:
          type: string
          pattern: '\S[@]\S'
    get:
      summary: Get user
      operationId: GetUser
      responses:
        '200':
          description: Get user successfully
          content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Response'
    put:
      summary: Update user
      operationId: UpdateUser
      requestBody: 
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        '201':
          description: Update user successfully
          content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Response'
    delete:
      summary: Delete user
      operationId: UpdateUser
      parameters:
        - name: jwt
          in: cookie
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Delete user successfully
          content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Response'
...
components:
  schemas:
    Response:
      type: object
      properties:
        code:
          type: integer
          format: int32
        description:
          type: string
...
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          pattern: '\S[@]\S'
        password:
          type: string
...